tests: add special test setup where "gc" is run
It's very useful to smoke out concurrency and locking issues in the GC
code by using the entire tests suite as a stress test. Add a
GIT_TEST_GC=true special test setup that fakes up need_to_gc() to
always return true. See the documentation changes in t/README for
details.

The GIT_TEST_GC=true mode is also useful for ad-hoc testing of and
debugging of "gc --auto", since it emulates a case where need_to_gc()
will return true. Setting gc.auto=0 would disable "gc --auto"
entirely, and setting gc.auto=1 would require manually injecting one
SHA-1 looking file into objects/17/ to trigger the heuristic.

There's still a lot of test failures under this mode, so marking up
everything with the "GC" prerequisite is a work-in-progress. But I
already find this useful enough, so let's include this even though
it's not fully finished.

There's a few test changes to t/*.sh that give a good idea of the sort
of work we need to do to get the test suite to pass. Some of this
overlaps with the reftable effort (or other ref backend
abstraction). I.e. some tests that aren't interested in stressing the
ref backend itself are needlessly peeking behind the curtain to check
ref values when they should use e.g. "rev-parse". Other tests will
legitimately need to disable this because they're e.g. testing
repacking or expiry, and running "gc" after the common ref-modifying
porcelain commands will fail those test.

In the future we could do something with GIT_TEST_GC=<true> and
GIT_TEST_COMMIT_GRAPH=<true> in combination. Now the latter only
generates the graph on "commit", perhaps we'd smoke out more issues if
we also generated it in the other places that "gc --auto".

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>