gc: remove redundant is_tempfile_active() code
Remove redundant code originally added in 4c5baf0273 ("gc: remove
gc.pid file at end of execution", 2013-10-16). Back then the pidfile
was manually managed, and this was an "if (pidfile)" test.

That variable would only be set in cmd_gc(), and back then there was
no gc.autoDetach=true, so this wouldn't e.g. be set in a parent and
the lock retried in the child.

So there was never any reason to do this, it amounts to a check of
whether a variable we know we set once is set twice, and if the code
was ever changed to introduce such a bug we'd die a few lines later in
hold_lock_file_for_update() which'll be passed "LOCK_DIE_ON_ERROR" and
die if there's an attempt to re-lock the already-locked file, even
within the same process.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>