remote-mediawiki tests: use CLI installer
Replace the use of screen-scraping in the test environment
installation with simply invoking MediaWiki's command-line
installer.

The old code being deleted here relied on our own hardcoded POST
parameter names & the precise layout of MediaWiki's GUI installer at a
given version. Somewhere between [1] and now this inevitably
broke. Let's simply use MediaWiki's CLI installation mode instead.

As far as I can tell there was never a reason for this hack, when [1]
was introduced it hardcoded MediaWiki 1.19.0, the CLI installer was
introduced in 1.17.0. Perhaps the authors weren't aware of it, or this
code was written for an older version.

This allows us to simply delete our own template version of
LocalSettings.php, it'll instead be provided by the CLI
installer.

While we're at it let's fix a few things, these changes weren't
practical to split up (I'd need to fix code I was about to mostly
delete):

 * Use MediaWiki's own defaults for simplicity, e.g. before we'd name
   the database "wikidb.sqlite", now we'll simply use whatever name
   MediaWiki prefers by only supplying the directory name the SQLite
   file will be dropped into, not the full path.

  * Put all of our database & download assets into a new "mediawiki/"
    folder. This makes it easier to reason about as the current &
    template "backup" database the tests keep swapping around live
    next to each other.

  * The "delete" function now only deletes the MediaWiki installation
    & database, not the downloaded .tar.gz file. This makes us
    friendlier to a developer on a slow connection.

1. 5ef6ad1785 ("git-remote-mediawiki: scripts to install, delete and
   clear a MediaWiki", 2012-07-06)

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>