hook: allow parallel hook execution
In many cases, there's no reason not to allow hooks to execute in
parallel. run_processes_parallel() is well-suited - it's a task queue
that runs its housekeeping in series, which means users don't
need to worry about thread safety on their callback data.

So let's blacklist those those hooks that require a single job in the
common case (e.g. because they're expected to edit a file in-place)
with a new "RUN_HOOKS_OPT_INIT_SYNC" initializer, and make parallel
execution the default.

Teach run_hooks() to use that function for simple hooks which don't
require stdin or capture of stderr.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>
Helped-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>