cook: add github.com/git-vcs/git as a URL
I maintain this mirror, but hopefully we can make it semi-official. It
has the same refs as the GitHub one except for the GitHub "pull" refs:

    $ diff -u \
        <(git ls-remote https://gitlab.com/git-vcs/git.git/) \
        <(git ls-remote https://github.com/git/git/|grep -v refs/pull)
    $

Although I had to delete the stale "pu" manually just now.

As an aside there are also https://gitlab.com/gitlab-org/git and
https://gitlab.com/git-vcs/git which mostly mirror but also carry some
GitLab Employee topic branches.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>