pretty format %(trailers): avoid needless repetition
Change the documentation for the various %(trailers) options so it
isn't repeating part of the documentation for "only" about how boolean
values are handled. Instead let's split the description of that into
general documentation at the top.

It then suffices to refer to it by listing the options as
"opt[=bool]". I'm also changing it to "[=bool]" from "[=val]". It took
me a couple of readings to realize that while to realize that these
options were referring back to the "only" option's treatment of
boolean values. Let's try to make this more explicit.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>